Fix bug: Backend servers muddle requests from different clients
Since the LoadBalancer forwarded requests from different clients on single channels to the backend servers, the servers saw only one client on their side and mixed request data from different clients in to one, which resulted in errors.

Additionally, make the TCPClient aware of input data, which may contain the separator sign, which belongs to the protocol's semantics. So the client will remove all occurrences of the separator.